♻️ Refactor: Separate CreateCharacter screen and add tab navigation logic
- CreateCharacter 컴포넌트를 CreateCharacterScreen 디렉토리로 이동하여 파일 구조 정리
- 탭 클릭 시 nowScreen 상태에 따라 각 화면이 구분되어 보이도록 탭 분기 로직 추가
- 현재 Content 컴포넌트로 탭 화면 연결 (추후 각 탭에 맞는 컴포넌트로 교체 예정)
